requirements.txt
alembic==1.4.3
aniso8601==8.0.0
appdirs==1.4.4
arrow==0.17.0
# astroid==2.4.2
attrs==17.4.0
autopep8==1.5.5
beautifulsoup4==4.9.3
behave==1.2.6
binaryornot==0.4.4
blinker==1.4
boto==2.49.0
boto3==1.16.61
python-dotenv==0.15.0
# botocore==1.19.62
# browserstack-local==1.2.3
# bump2version==1.0.1
certifi==2020.11.8
cffi==1.14.4
cfn-lint==0.44.7
chardet==3.0.4
click==7.1.2
colorama==0.4.3
# configparser==5.0.1
cookiecutter==1.7.2
crayons==0.4.0
cryptography==3.3.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
distlib==0.3.1
docopt==0.6.2
docutils==0.15.2
ecdsa==0.14.1
extras==1.0.0
filelock==3.0.12
fixtures==3.0.0
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.9
flask-heroku==0.1.9
Flask-Jsonpify==1.5.0
Flask-Login==0.5.0
flask-media==0.4.6
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
Flask-WTF==0.14.3
future==0.18.2
fuzzywuzzy==0.18.0
geoffrey-csslint==0.0.1
gunicorn==19.10.0
heroku3==4.2.3
html-linter==0.4.0
idna==2.10
importlib-metadata==2.0.0
imutils==0.5.4
iniconfig==1.1.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsmin==2.2.2
jsondiff==1.2.0
jsonpatch==1.28
jsonpickle==2.0.0
jsonpointer==2.0
jsonschema==3.2.0
junit-xml==1.9
lazy-object-proxy==1.4.3
linecache2==1.0.0
Mako==1.1.3
matplotlib==2.2.5
mccabe==0.6.1
moto==1.3.16
msrest==0.6.21
numpy
oauthlib==3.1.0
opencv-python
opencv-python-headless==4.4.0.46
openstep-parser==1.5.3
packaging==20.4
pandas==1.2.0
parse==1.19.0
parse-type==0.5.2
parsy==1.1.0
pathlib2==2.3.5
pathspec==0.8.1
pbr==5.5.1
pbxproj==3.2.0
pip-check-reqs==2.1.1
plotly==4.14.3
pluggy==0.13.1
poyo==0.5.0
psutil==5.8.0
psycopg2-binary==2.8.6
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pyflakes==2.2.0
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-mimeparse==1.6.0
python-slugify==4.0.1
python-subunit==1.4.0
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
responses==0.12.1
retrying==1.3.3
rsa==4.5
s3transfer==0.3.4
selenium==3.141.0
sh==1.14.1
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.1
SQLAlchemy==1.3.20
sshpubkeys==3.3.1
sklearn
sure==1.4.11
template-remover==0.1.9
termcolor==1.1.0
testtools==2.4.0
text-unidecode==1.3
toml==0.10.1
tox==3.21.1
traceback2==1.4.0
typed-ast==1.4.1
unittest2==1.1.0
urllib3==1.26.2
virtualenv==20.1.0
waitress==1.4.4
Wand==0.6.5
webdriver-manager==3.3.0
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xmltodict==0.12.0
# yamllint==1.25.0
zipp==3.4.0
chromedriver-py==89.0.4389.23
